Update app.json
The main point of this commit is to change the default `ABE_URL` value to `https://abe-dev.olin.build`, so that newly-minted PR apps point to dev.

Also took the opportunity to add `repository` and `success_url` properties as documented here <https://devcenter.heroku.com/articles/app-json-schema>, and remove some of the optional empty values.