Add a Labels page, for admins to review labels
Adds a  Labels page, at /labels. This page makes it easier for an administrator to review the labels, their descriptions, colors, and flags.

This page isn't part of the public UI. It's not protected by auth, since it doesn't (yet) allow modifications to the set of labels, but there's no link to it.