Hide the Add and Edit buttons from unauthenticated users
Start #217, #179.

Remaining work depends on olinlibrary/ABE#214. This work is to retrieve the account info from GET /account instead of initializing it to a constant that grants all permissions; to extend withServerInfo (currently in PR #219) to verify that the account has been loaded; and to guard the Sidebar by withServerInfo.